soz
channelId is broken.
Someone was in the process of trying to get rid of it and replace it with a function to callback (which looks much better), but I was using channelId for the top function too, which needs it because we have to put the datepicker out (not just {channel:, text:}). Therefore, we cannot use the callback there (without either getting bacj channelId or changing callback? idk)

Thank god for history. We can review in the morning. I think the git merge was awful from me...